+++ database/migrations/tenant/2022_09_02_064359_create_shop_stats.php
@@ -16,45 +16,46 @@ return new class () extends Migration {
             $table->smallIncrements('id');
             $table->unsignedSmallInteger('shop_id')->index();
             $table->foreign('shop_id')->references('id')->on('shops');
-            $table->unsignedBigInteger('number_customers')->default(0);
+
+            $table->unsignedInteger('number_customers')->default(0);
             $customerStates = ['in-process', 'active', 'losing', 'lost', 'registered'];
             foreach ($customerStates as $customerState) {
-                $table->unsignedBigInteger('number_customers_state_'.str_replace('-', '_', $customerState))->default(0);
+                $table->unsignedInteger('number_customers_state_'.str_replace('-', '_', $customerState))->default(0);
             }
             $customerNumberInvoicesStates = ['none', 'one', 'many'];
             foreach ($customerNumberInvoicesStates as $customerNumberInvoicesState) {
-                $table->unsignedBigInteger('number_customers_trade_state_'.$customerNumberInvoicesState)->default(0);
+                $table->unsignedInteger('number_customers_trade_state_'.$customerNumberInvoicesState)->default(0);
             }
 
-            $table->unsignedBigInteger('number_departments')->default(0);
+            $table->unsignedInteger('number_departments')->default(0);
             $departmentStates = ['in-process', 'active', 'discontinuing', 'discontinued'];
             foreach ($departmentStates as $departmentState) {
-                $table->unsignedBigInteger('number_departments_state_'.str_replace('-', '_', $departmentState))->default(0);
+                $table->unsignedInteger('number_departments_state_'.str_replace('-', '_', $departmentState))->default(0);
             }
 
-            $table->unsignedBigInteger('number_families')->default(0);
+            $table->unsignedInteger('number_families')->default(0);
             $familyStates = ['in-process', 'active', 'discontinuing', 'discontinued'];
             foreach ($familyStates as $familyState) {
-                $table->unsignedBigInteger('number_families_state_'.str_replace('-', '_', $familyState))->default(0);
+                $table->unsignedInteger('number_families_state_'.str_replace('-', '_', $familyState))->default(0);
             }
-            $table->unsignedBigInteger('number_orphan_families')->default(0);
+            $table->unsignedInteger('number_orphan_families')->default(0);
 
-            $table->unsignedBigInteger('number_products')->default(0);
+            $table->unsignedInteger('number_products')->default(0);
             $productStates = ['in-process', 'active', 'discontinuing', 'discontinued'];
             foreach ($productStates as $productState) {
-                $table->unsignedBigInteger('number_products_state_'.str_replace('-', '_', $productState))->default(0);
+                $table->unsignedInteger('number_products_state_'.str_replace('-', '_', $productState))->default(0);
             }
 
 
-            $table->unsignedBigInteger('number_orders')->default(0);
+            $table->unsignedInteger('number_orders')->default(0);
             $orderStates = ['in-basket', 'in-process', 'in-warehouse', 'packed', 'finalised', 'dispatched', 'returned', 'cancelled'];
             foreach ($orderStates as $orderState) {
-                $table->unsignedBigInteger('number_orders_state_'.str_replace('-', '_', $orderState))->default(0);
+                $table->unsignedInteger('number_orders_state_'.str_replace('-', '_', $orderState))->default(0);
             }
 
-            $table->unsignedBigInteger('number_deliveries')->default(0);
-            $table->unsignedBigInteger('number_deliveries_type_order')->default(0);
-            $table->unsignedBigInteger('number_deliveries_type_replacement')->default(0);
+            $table->unsignedInteger('number_deliveries')->default(0);
+            $table->unsignedInteger('number_deliveries_type_order')->default(0);
+            $table->unsignedInteger('number_deliveries_type_replacement')->default(0);
 
 
             $deliveryStates = [
@@ -70,20 +71,20 @@ return new class () extends Migration {
             ];
 
             foreach ($deliveryStates as $deliveryState) {
-                $table->unsignedBigInteger('number_deliveries_state_'.str_replace('-', '_', $deliveryState))->default(0);
+                $table->unsignedInteger('number_deliveries_state_'.str_replace('-', '_', $deliveryState))->default(0);
             }
 
             foreach ($deliveryStates as $deliveryState) {
-                $table->unsignedBigInteger('number_deliveries_cancelled_at_state_'.str_replace('-', '_', $deliveryState))->default(0);
+                $table->unsignedInteger('number_deliveries_cancelled_at_state_'.str_replace('-', '_', $deliveryState))->default(0);
             }
 
-            $table->unsignedBigInteger('number_invoices')->default(0);
-            $table->unsignedBigInteger('number_invoices_type_invoice')->default(0);
-            $table->unsignedBigInteger('number_invoices_type_refund')->default(0);
+            $table->unsignedInteger('number_invoices')->default(0);
+            $table->unsignedInteger('number_invoices_type_invoice')->default(0);
+            $table->unsignedInteger('number_invoices_type_refund')->default(0);
 
-            $table->unsignedBigInteger('number_payment_service_providers')->default(0);
-            $table->unsignedBigInteger('number_payment_accounts')->default(0);
-            $table->unsignedBigInteger('number_payments')->default(0);
+            $table->unsignedInteger('number_payment_service_providers')->default(0);
+            $table->unsignedInteger('number_payment_accounts')->default(0);
+            $table->unsignedInteger('number_payments')->default(0);
 
 
             $table->timestampsTz();
